% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{accuracy_tbl}
\alias{accuracy_tbl}
\title{Accuracy table (MAE, RMSE, MAPE, sMAPE, MASE, Theil_U, R2)}
\usage{
accuracy_tbl(act, pred)
}
\arguments{
\item{act}{numeric actual values}

\item{pred}{numeric predicted values}
}
\value{
data.frame with metrics
}
\description{
Accuracy table (MAE, RMSE, MAPE, sMAPE, MASE, Theil_U, R2)
}
