% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xreg.R
\name{build_xreg_main}
\alias{build_xreg_main}
\title{Build main xreg (y_t ~ x_{t-L}) for forward forecasting}
\usage{
build_xreg_main(y_ts, ex_ts, L, h_out)
}
\arguments{
\item{y_ts}{ts}

\item{ex_ts}{ts}

\item{L}{lag}

\item{h_out}{horizon}
}
\value{
list(x_tr, x_te)
}
\description{
Build main xreg (y_t ~ x_{t-L}) for forward forecasting
}
