% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xreg.R
\name{forecast_exog}
\alias{forecast_exog}
\title{Forecast exogenous series with ARIMA}
\usage{
forecast_exog(series, h)
}
\arguments{
\item{series}{numeric or ts}

\item{h}{horizon}
}
\value{
ts
}
\description{
Forecast exogenous series with ARIMA
}
