% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif.R
\name{make_forecast_gif}
\alias{make_forecast_gif}
\title{Create animation (Actual + Backtest Forecast + Future Forecast)}
\usage{
make_forecast_gif(
  combined,
  pg_name,
  bt_df = NULL,
  backtest_start_date = as.Date("2023-01-01"),
  duration = 8,
  fps = 20,
  width = 900,
  height = 500
)
}
\arguments{
\item{combined}{data.frame(date,value,type)}

\item{pg_name}{label}

\item{bt_df}{optional backtest df(date,Actual,Forecast)}

\item{backtest_start_date}{Date}

\item{duration}{seconds}

\item{fps}{frames per second}

\item{width}{px}

\item{height}{px}
}
\value{
list(gif, preview)
}
\description{
Create animation (Actual + Backtest Forecast + Future Forecast)
}
