% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{plot_backtest}
\alias{plot_backtest}
\title{Plot backtest Actual vs Forecast}
\usage{
plot_backtest(bt_df, pg_name, file_name)
}
\arguments{
\item{bt_df}{data.frame}

\item{pg_name}{label}

\item{file_name}{png path}
}
\value{
invisible NULL
}
\description{
Plot backtest Actual vs Forecast
}
