% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{rolling_cv_rmse}
\alias{rolling_cv_rmse}
\title{Rolling-origin CV RMSE}
\usage{
rolling_cv_rmse(
  y_ts,
  xreg,
  init = 48,
  h = 12,
  thr_fail = 0.1,
  stepwise = TRUE,
  approximation = TRUE
)
}
\arguments{
\item{y_ts}{ts}

\item{xreg}{matrix}

\item{init}{initial window size}

\item{h}{horizon}

\item{thr_fail}{failure ratio to switch fallback}

\item{stepwise}{logical}

\item{approximation}{logical}
}
\value{
numeric RMSE
}
\description{
Rolling-origin CV RMSE
}
