% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{run_backtest_from}
\alias{run_backtest_from}
\title{Run leak-free backtest from a given year-month}
\usage{
run_backtest_from(
  y_ts,
  ex_ts,
  shift,
  start_year = 2023,
  start_month = 1,
  L = 1
)
}
\arguments{
\item{y_ts}{ts (log scale)}

\item{ex_ts}{ts (exogenous)}

\item{shift}{numeric (for invlog)}

\item{start_year}{integer}

\item{start_month}{integer}

\item{L}{lag}
}
\value{
data.frame(date, Actual, Forecast)
}
\description{
Run leak-free backtest from a given year-month
}
