% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{run_pca}
\alias{run_pca}
\title{PCA on CSI1:CSI85 and scree plot}
\usage{
run_pca(dt, csi_cols, pg_name)
}
\arguments{
\item{dt}{data.table}

\item{csi_cols}{character}

\item{pg_name}{label}
}
\value{
list(dt, pc_names, loadings, scree_file, scree_df)
}
\description{
PCA on CSI1:CSI85 and scree plot
}
